+++ arcana/core/data/tests/test_column.py
@@ -1,7 +1,7 @@
 import pytest
 from operator import mul
 from functools import reduce
-from arcana.core.data.item import FileGroup
+from arcana.core.data.format import FileGroup
 
 @pytest.mark.skip("needs to wait until further refactoring")
 def test_column_api_access(dataset):
