+++ arcana/data/formats/common.py
@@ -1,34 +1,134 @@
-from arcana.core.data.format import FileFormat
+from abc import ABCMeta
+from arcana.core.data.format import FileGroup, BaseFile, BaseDirectory
+from arcana.core.mark import converter
 from arcana.tasks.archive import (
     create_tar, extract_tar, create_zip, extract_zip)
 
 
 # General formats
-directory = FileFormat(name='directory', extension=None, directory=True)
-text = FileFormat(name='text', extension='.txt')
-json = FileFormat(name='json', extension='.json')
+class Text(BaseFile):
+    ext = 'txt'
+
+class Csv(BaseFile):
+    ext = 'csv'
+
+class Tsv(BaseFile):
+    ext = 'tsv'
+
+class TextMatrix(BaseFile):
+    ext = 'mat'
+
+class RFile(BaseFile):
+    ext = 'rData'
+
+class MatlabMatrix(BaseFile):
+    ext = 'mat'
+
 
 # Compressed formats
-zip = FileFormat(name='zip', extension='.zip')
-tar = FileFormat(name='targz', extension='.tar')
-targz = FileFormat(name='targz', extension='.tar.gz')
+class Zip(BaseFile):
+    ext = 'zip'
 
-standard_formats = [text, json, directory, zip, targz]
+    @classmethod
+    @converter(FileGroup)
+    def archive(cls, file_group, **kwargs):
+        node = create_zip(
+            in_file=file_group,
+            compression='gz',
+            **kwargs)
+        return node, node.lzout.out_file
+
+class Gzip(BaseFile):
+    ext = 'gz'
+
+    @classmethod
+    @converter(FileGroup)
+    def archive(cls, file_group):
+        raise NotImplementedError
+
+class Tar(BaseFile):
+    ext = 'tar'
+
+    @classmethod
+    @converter(FileGroup)
+    def archive(cls, file_group, **kwargs):
+        node = create_tar(
+            in_file=file_group,
+            compression='gz',
+            **kwargs)
+        return node.lzout.out_file    
+
+class TarGz(Tar, Gzip):
+    ext = 'tar.gz'
+
+    @classmethod
+    @converter(FileGroup)
+    def archive(cls, file_group, **kwargs):
+        node = create_tar(
+            in_file=file_group,
+            compression='gz',
+            **kwargs)
+        return node.lzout.out_file
+
+
+class Directory(BaseDirectory):
+
+    @classmethod
+    @converter(Zip)
+    def unzip(cls, zip_file, **kwargs):
+        node = extract_zip(
+            in_file=zip_file,
+            **kwargs)
+        return node.lzout.out_file
+
+    @classmethod
+    @converter(Tar)
+    def untar(cls, tar_file, **kwargs):
+        node = extract_tar(
+            in_file=tar_file,
+            **kwargs)
+        return node.lzout.out_file
+
+    @classmethod
+    @converter(TarGz)
+    def untargz(cls, tar_file, **kwargs):
+        node = extract_tar(
+            in_file=tar_file,
+            **kwargs)
+        return node.lzout.out_file
+
+
+# Hierarchical text files
+
+class HierarchicalText(BaseFile):
+    pass
+
+class Json(HierarchicalText):
+    ext = 'json'
+
+class Yaml(HierarchicalText):
+    ext = 'yml'    
+
+
+standard_formats = [Text, Directory, Zip, Tar, TarGz]
 
 # General image formats
-gif = FileFormat(name='gif', extension='.gif')
-png = FileFormat(name='png', extension='.png')
-jpg = FileFormat(name='jpg', extension='.jpg')
+class ImageFile(BaseFile, metaclass=ABCMeta):
+    pass
+
+class Gif(ImageFile):
+    ext = 'gif'
+
+class Png(ImageFile):
+    ext = 'png'
+
+class Jpeg(ImageFile):
+    ext = 'jpg'
+
 
 # Document formats
-pdf_format = FileFormat(name='pdf', extension='.pdf')
+class Document(BaseFile, metaclass=ABCMeta):
+    pass
 
-# Set Converters
-directory.set_converter(zip, extract_zip)
-directory.set_converter(targz, extract_tar)
-directory.set_converter(tar, extract_tar)
-tar.set_converter(directory, create_tar)
-targz.set_converter(directory, create_tar, compression='gz')
-zip.set_converter(directory, create_zip)
-text.set_converter(zip, extract_zip)
-zip.set_converter(text, create_zip)
+class Pdf(Document):
+    ext = 'pdf'
